Fix bug for Scalable-Broadcast
When running the current Scalable-Broadcast demo (https://rtcmulticonnection.herokuapp.com/demos/Scalable-Broadcast.html) watchers on at least chrome and firefox cannot open a stream. This is because the browser tries to set the local answer sdp in a stable state. The solution is to not negotiate a new state when the state is stable.